Address PR #21 review feedback: blocking fixes, canonical filter spec, structural cleanup
Blocking:
- Fix analyze.py CSV crash when lines_with_bbox present but img dimensions are 0
- Remove dead amplitude key from moire config (never read by MoireOverlayEffect)

Structural:
- Add QUALITY_FILTER_DEFAULTS to serialization.py as canonical filter threshold spec
- analyze.py: replace _REPORT_FILTER_THRESHOLDS with QUALITY_FILTER_DEFAULTS import;
  update aggregate_stats problematic_samples to use canonical thresholds; remove
  unused workers param from _build_report/_cmd_report/report subparser
- template.py: use QUALITY_FILTER_DEFAULTS for all 6 filter attribute defaults
- elements/content.py: extract _render_cells helper, eliminating ~25-line duplication
  between _render_zone and the body GridStack loop
- Add IoU vs containment-fraction semantic docstrings to calculate_iou / analyze_line_overlaps

Small fixes:
- effects/physical.py: move PIL imports to module top; narrow except Exception → OSError
- template.py: remove redundant os.makedirs (output_dirpath already created above)
- annotations.py: cast gray to float32 (halves memory for Laplacian computation)
- SYNTHDOG-VALIDATION.md: add Thread 11 documenting elastic distortion n=50 study
- .pre-commit-config.yaml: lower --maxkb from 2500 to 1000

+++ data/synthdog_grounding/annotations.py
@@ -200,7 +200,7 @@ def compute_quality_metrics(
     total_ct: int,
 ) -> dict:
     """Compute per-sample quality metrics from the rendered image."""
-    gray = 0.2989 * image[..., 0] + 0.5870 * image[..., 1] + 0.1140 * image[..., 2]
+    gray = (0.2989 * image[..., 0] + 0.5870 * image[..., 1] + 0.1140 * image[..., 2]).astype(np.float32)
     line_contrasts = []
     line_contrast_ratios = []
     line_bbox_areas_px = []
